refactor(domain): move query-param types out of endpoints layer
