+++ src/ksef2/endpoints/async_peppol.py
@@ -3,8 +3,8 @@
 from typing import Unpack, final
 
 from ksef2.core import routes
+from ksef2.domain.types import OffsetPaginationQueryParams
 from ksef2.endpoints.async_base import AsyncBaseEndpoints
-from ksef2.endpoints.base import OffsetPaginationQueryParams
 from ksef2.infra.schema.api import spec
