+++ src/ksef2/endpoints/async_tokens.py
@@ -1,23 +1,14 @@
 """Async token endpoints for managing access tokens."""
 
-from typing import NotRequired, TypedDict, Unpack, final
+from typing import Unpack, final
 
 from pydantic import TypeAdapter
 
 from ksef2.core import routes
+from ksef2.domain.types import ListTokensQueryParams
 from ksef2.endpoints.async_base import AsyncBaseEndpoints
 from ksef2.infra.schema.api import spec
 
-ListTokensQueryParams = TypedDict(
-    "ListTokensQueryParams",
-    {
-        "status": NotRequired[list[str] | None],
-        "description": NotRequired[str | None],
-        "authorIdentifier": NotRequired[str | None],
-        "authorIdentifierType": NotRequired[str | None],
-        "pageSize": NotRequired[int | None],
-    },
-)
 _LIST_TOKENS_PARAMS = TypeAdapter(ListTokensQueryParams)
